input_extraction/requirements.txt
google-cloud-storage==2.18.2
pdfplumber==0.11.4
python-docx==1.1.2
pika==1.3.2
python-dotenv==1.0.1
